==> src/2. Try_data9.Rmd <==
---
title: "TRY database"
author: "Jelyn Gerkema"
date: "2023-11-17"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 0. Set up
```{r set up}
source("src/settings.R")
source("src/functions.R")

#try_categorical <- read_excel("data/raw/TRY/TRY_Categorical_Traits_Lookup_Table_2012_03_17_TestRelease.xlsx")

#try_taxonomic <- read_excel("data/raw/TRY/try_taxonomic.xlsx")

#library(WorldFlora)

classification <- read_delim("data/raw/classification202312.csv", 
    delim = "\t", escape_double = FALSE, 
    trim_ws = TRUE)

#library(rtry)

```


```{r try}

#temp <- TRYdata1 %>%   filter(AccSpeciesName %in% c(synonyms$scientificName)) %>%   select(AccSpeciesName, SpeciesName) %>%   distinct()

# synonyms <- classification %>%   filter(taxonomicStatus == "Synonym") 

length(unique(TRYdata1$AccSpeciesName)) # n = 94230

path_to_data <- "data/raw/TRY/29649.txt"

TRYdata1 <- rtry_import(path_to_data)


#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 413)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)

TRYdata2 <- TRYdata1 %>%
  filter(TraitID %notin% c(3107, 18))

length(unique(TRYdata2$TraitName)) # n = 94230


try_ls <- rtry_exclude(TRYdata1, (DataID %in% 413) & 
                         (OrigValueStr %in% c("13 weeks after sowing, 8 weeks after start of treatments",
          "check", "compound", "immature", "intermed", "junvenil", "juvenil", "Juvenil", "juvenil (3 years)",
          "juvenil: 0.5-2.5 m high", "juvenile", "Juvenile", "juvenile, 11-14 weeks", "Juvenile, 15 years",
          "juvenile, 6 weeks", "S", "sapling", "Sapling (1 - 10 y)", "saplings", "saplings (height 1-2m)",
          "saplings, 5 - 10 years", "second years after sowing seeds (relatively juvenile for most of species)",
          "seedling", "Seedling", "Seedling (0 - 1 y)", "seedlings", "seedlings, < 1/2 year", 
          "seedlings, 1st year", "seedlings, 6 month old", "T", "T/F", "T?", "Y", "young", "Young leaves", 
          "young tree", "young trees at  0.5-3m height", "young, not seedlings")), baseOn = ObservationID)


#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 327)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)

try_ls_exp <- rtry_exclude(try_ls, (DataID %in% 327) & 
                         (OrigValueStr %in% c("BAU-Abandonnée/Nardaie Reposoir", 
          "BAU-Abandonnée/Nardaie non paturée", "BAU-Abandonnée/prairie grasse non paturée", 
          "BAU-Abandonée/Seslérie non paturée.", "BAU-Nardaie paturée.", 
          "BAU-Paturée Intermediaire Grasse-seslérie",
          "BAU-Prairie grasse paturée", "BAU-Prairie à Seslérie paturée", "Branch Bag", "C", "CG", "Chamber",
          "Climate Chamber", "Climate chamber, non-limiting conditions, (cf. dataset reference)", 
          "Common Garden", "Common Garden Experiment", "Common garden", "Common garden trees", 
          "Controlled climate chamber", "Dry Plot", "Experimental, Glasshouse", "FACE",
          "FE", "Field Experiment", "Fully open overstory 90 days, seedling", 
          "Fully open overstory, seedling", "G", "GH", "Glass house", "Glasshouse", 
          "Glasshouse experiment", "Greehouse", "Green house", "Greenhouse", "Greenhouse plants",
          "Greenhouse, Indiana University", "Greenhouse, grrowth container", 
          "Greenhouse: highlight_highpH_competition", "Greenhouse: highlight_highpH_nocompetition",
          "Greenhouse: highlight_lowpH_competition", "Greenhouse: highlight_lowpH_nocompetition",
          "Greenhouse: lowleight_lowpH_competition", "Greenhouse: lowlight_highpH_competition",
          "Greenhouse: lowlight_highpH_nocompetition", "Greenhouse: lowlight_lowpH_nocompetition",
          "Grown from seed sown in outdoor field experiment", "Growth Chamber", "Growth chamber", 
          "Growth exp", "Harte warming experiment (montane meadow)", 
          "I do not like this paper. Results looks questionable", 
          "Irrigation and N fertilisation (100 kg/ha)", "LAU_Grazed/Abandoned", 
          "LAU_Never ploughed/grazed", "LAU_Never ploughed/grazed highland", 
          "LAU_Never ploughed/grazed slopes", "LAU_Never ploughed/mown", "LAU_Ploughed/grazed",
          "LAU_Ploughed/mown", "LAU_Ploughed/mown and fertilized", 
          "Large potted monoculture mesocosms, outside, soil from local meadow,  grazing and harvests emulated",
          "Mature cultivated", "Monoculture", "Natural/C", "OTC", "OTC O3", "OTC field", 
          "Open Top", "Open top chambers", "Other", "Outdoor cultivation in 30x30cm pots, standardized substrate",
          "Outdoor cultivations (4 repititions) grown in 30x30cm pots with standardized substrate. All pot ...",
          "PM", "PU", "Partially open overstory 90 days, seedling", "Planted trees", "Planted vines",
          "Pot exp", "Pot-grown", "Pots outside", "Siemianice", "Undisturbed, seedling", 
          "University of California Botanical Garden", "VER_permanent meadow mown and fertilized",
          "ER_permanent meadows mown and fertilized", "WTC", "botanical garden environment",
          "branch bag", "climate chamber", "climate chambers", "comm.garden.Spain", 
          "common garden in growth containers with soil corresponding to seed origin", 
          "common garden, understorey", "controlled environment", "controlled environment room", 
          "cultivated", "disturbed soil treatment; coniferous forest", 
          "disturbed soil treatment; deciduous alluvial forest", "disturbed soil treatment; fallow wet meadow",
          "drought treatment", "dry season measurements (see also #659)", "experimental",
          "experimental treatment", 
          "experimental warming with open top chambers during the snow-free period, increasing daily mean  ...",
          "experimental, in pots outside", "experimental, outside in pots", "field experiment",
          "forest fertilization", "glasshouse, climate controlled", 
          "glasshouse, plastic tubes 10cm diameter, 100 cm depth", "grassland common garden", "greenhouse",
          "greenhouse - Northern Arizona University (standard controlled environment)", "growth chamber",
          "growth-chamber", "growth_chamber", "hydroponic", "lab", "mesocosm", "mesocosm, 50l pots",
          "mini-ecosystem", "natural environment, high warming +4C, preccipitation ambient",
          "natural environment, high warming +4C, preccipitation ambient +50%", 
          "natural environment, high warming +4C, preccipitation ambient -50%",
          "natural environment, low warming +1.5C, preccipitation ambient",
          "natural environment, low warming +1.5C, preccipitation ambient +50%", 
          "natural environment, low warming +1.5C, preccipitation ambient -50%", 
          "natural environment, medium warming +2.5C, preccipitation ambient",
          "natural environment, medium warming +2.5C, preccipitation ambient +50%",
          "natural environment, medium warming +2.5C, preccipitation ambient -50%",
          "natural environment, no warming, preccipitation ambient +50%",
          "natural environment, no warming, preccipitation ambient -50%",
          "natural grassland, experimental nutrient NP addition", "nursery",
          "nutrient addition experiment", "open-sided growth chamber", "open-top chamber",
          "open-top chambers", "pot", "pots, outside in natural environment", "shade houses",
          "smelter polluted environment", "water stress experiment", "water treatment", 
          "whole-tree chamber", "plantation")), baseOn = ObservationID)

#tmp_unfiltered <- rtry_select_row(TRYdata1, DataID %in% 1961)

#tmp_unfiltered <- rtry_explore(tmp_unfiltered, DataID, DataName, OriglName, OrigValueStr, OrigUnitStr, StdValue, Comment, sortBy = OrigValueStr)

try_ls_exp_hlth <- rtry_exclude(try_ls_exp, (DataID %in% 1961) & 
                         (OrigValueStr %in% c("Dead")), baseOn = ObservationID)

#try_covariate <- try_filtered %>%
 # filter(is.na(TraitID))


#covariates <- unique(try_covariate$DataName) %>% as.data.frame()

try_filtered <- rtry_remove_dup(try_ls_exp_hlth)

#tmp_unfiltered <- rtry_explore(try_filtered, DataID, DataName, TraitID, TraitName, ErrorRisk, sortBy = ErrorRisk)

try_filtered <- rtry_exclude(try_filtered, ErrorRisk >= 3, baseOn = ObsDataID) 



trait_explore <- rtry_explore(try_filtered, DataID, DataName, TraitID, TraitName, sortBy = TraitID)

try_clean <- try_filtered %>%
  #mutate(OrigValueStr =  case_when(OrigValueStr == "c3" ~ "C3",
   #                          OrigValueStr == "C3?" ~ "C3",
    #                         OrigValueStr == "c4" ~ "C4",
     #                        OrigValueStr == "C4?" ~ "C4",
      #                       TRUE ~ OrigValueStr)) %>%
  filter(DataID %notin% c(8178, 9780, 1629, 1739, 974, 2526, 2760, 2527, 2761, 
                          9906, 2762, 9578, 9577, 9576, 9575, 9574, 3710, 2958, 
                          9447, 8651, 8643, 8648, 8642, 8650, 8647, 8646, 8649, 
                          8645, 8644, 2373, 2380, 2365, 2367, 2375, 2381, 3809, 
                          3812, 3814, 3811, 3816, 225, 2082, 2142, 3990, 3991, 
                          2412, 2413, 2414, 904, 1271, 9926, 9931, 27, 9898, 
                          9899, 9900, 9901, 476, 2506, 829, 2499, 2500, 1027, 
                          1028, 2311, 4142, 3974, 3979, 3980, 3978, 3976, 3975, 
                          3977)) %>%
  filter(TraitID %in% c(3, 4, 13, 14, 15, 22, 24, 26, 27, 46, 47, 48, 53, 54, 
                        59, 70, 71, 83, 89, 145, 146, 164, 246, 334, 785, 819, 839, 
                        1047, 1080, 1181, 1256, 3106, 3113, 3117, 3588, 3801)) %>% #n = 36
  drop_na(StdValue) %>%
  select(AccSpeciesName, TraitID, ObservationID, TraitName, StdValue,
         Reference, LastName, FirstName) %>%
  group_by(AccSpeciesName, ObservationID, TraitID, TraitName,
         Reference, LastName, FirstName) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup()


length(unique(try_clean$AccSpeciesName)) # n = 69312


renv::init()


removed_because_nonsensical <- try_clean %>% 
  filter(str_count(AccSpeciesName, "\\S+") == 1) %>%
  select(AccSpeciesName) %>%
  distinct() 

############################# Poging 2
try_clean2 <- try_clean %>%
  filter(str_count(AccSpeciesName, "\\S+") > 1) %>% 
  #121 AccSpeciesNames are removed, as they are not actual species names, but 
  # things like: "herbaceaus", "fern", or "ANANI". Their string only consists of one word.
  mutate(AccSpeciesName = tolower(AccSpeciesName)) %>%
  mutate(AccSpeciesName = paste0(tools::toTitleCase(word(AccSpeciesName, 1)), " ",
                                  word(AccSpeciesName, 2, -1))) 



load("~/0. PhD/0. Chapter_I/data/processed/try_clean29_1.RData")
ref <- as.data.frame(unique(try_clean2$Reference))

write_xlsx(ref, path = "docs/try_refs.xlsx")

length(unique(try_clean2$AccSpeciesName))

# try_clean2 contains 69191 unique, "accepted" species names, 1592 of those
# are nonsensical names. 
69191 - 1592 +121

# try_wf_clean2 contains the species names in the TRY database that have a direct
# match with the WFO backbone. n = 64617.
try_wf_clean_first <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(scientificName %in% c(try_clean2$AccSpeciesName))

length(unique(try_wf_clean_first$scientificName))

try_names_save_1 <- try_wf_clean_first %>%
  select(scientificName) %>%
  rename(old_try_name = scientificName) %>%
  mutate(new_try_name = old_try_name) %>%
  distinct()


# These species are not present in the WFO backbone. That could be because, 
# 1). the species is not correctly spelled, or 2). because it is not an actual species name. 
# Here, first some not actual species names are removed. 
not_included_in_wf <- try_clean2 %>%
  filter(AccSpeciesName %notin% c(try_wf_clean_first$scientificName)) %>%
  select(AccSpeciesName) %>%
  distinct() %>%
  separate(AccSpeciesName, into = c("genus", "sp"), extra = "merge", remove = F) 

length(unique(not_included_in_wf$AccSpeciesName)) #4574

removed_because_nonsensical_final <- not_included_in_wf %>%
  filter(sp %in% c("sp", "sp.", "abandoned area", "african shrubs", "alba preta",
                      "alluvial", "alluvial meadow", "araba tuua", "arc pp144",
                      "arc pp146", "average", "broad-leaved", "broadleaf", 
                      "catinga species", "cf.", "cff", "chocolate","cottage garden hybrid'",
                      "crop", "cultivar", "cultivars", "da", "da-mata", "da mata",
                      "da rosca", "da terra firme", "dark green", "darkcrasa", "de",
                      "deciduous forest", "deciduous trees", "evergreen trees",
                      "field annual", "findlay's blue'", "fisch.", "fitzgerald river",
                      "forest species", "form taxon 'filifolia'", "fuzzy", 
                      "gorky", "grande", "grass", "grasses", "grassland", 
                      "herbaceous", "hutch.", "hybrid", "l.", "laeve", "leaf", "leafea",
                      "limestone", "lookalike", "m.", "meadow", "meadow steppe",
                      "melkboom", "mill.", "mosaic", "numm.", "oak", "ob", "palm",
                      "para", "pasture", "preta", "preto", "preto folha grauda",
                      "preto folha miuda", "Rosette forb", "shrub", "shrubs and trees",
                      "sj", "skinny", "skinny2", "skinnyret", "small", "ss", "steppe",
                      "tallmorph", "terete-leaved", "terra", "tree", "tree co2", 
                      "tree field", "trees", "wedding day'"
                      ) |genus %in% c("Mountain", "may", "Lauraceae", "Southafrica", "Tropical")|AccSpeciesName %in% c("Paeonia 'anne rosse'", "Mid liana", 
                                  "Mata mata", "Mata mata preto", 
                                  "Mata mata si", "Buxuxu new", "Buxuxu old",
                                  "Sold pus", "Pubescent rosette", "Carex saw",
                                  "Eri sch", "Ucuuba t folha", "Death valley annual", 
                                  "Death valley annuals", "Deciduous chaparral shrubs"
                                  )) %>% # mata mata is a turtle species
  select(AccSpeciesName) %>%
  bind_rows(removed_because_nonsensical) %>%
  arrange(AccSpeciesName) 

length(unique(removed_because_nonsensical_final$AccSpeciesName)) #1592

not_included_in_wf <- not_included_in_wf %>%
  filter(AccSpeciesName %notin% c(removed_because_nonsensical_final$AccSpeciesName))

length(unique(not_included_in_wf$AccSpeciesName)) #3103

#write_xlsx(removed_because_nonsensical_final, path = "removed_because_nonsensical_final291.xlsx")


################################################## 
# A new clean try database is created, containing the original try names, and
# their new names which have a match in the WFO database. The original and new names
# are the same if the original directly matches with the wfo database. 

second_check_match <- not_included_in_wf %>%
  separate(AccSpeciesName, into = c("genus", "sp", "extra"), extra = "drop", remove = F) %>%
  filter(extra %in% c("var", "subsp", "x")) %>%
  mutate(scientificName = paste0(genus, " ", sp)) %>%
  select(-c("sp", "extra", "genus"))

length(unique(second_check_match$AccSpeciesName)) #n = 896
length(unique(second_check_match$scientificName)) #n = 866

try_wf_clean_second <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(scientificName %in% c(second_check_match$scientificName)) 

second_check_matched <- second_check_match %>%
  filter(scientificName %in% c(try_wf_clean_second$scientificName)) 

try_names_save_2 <- second_check_matched %>%
  rename(old_try_name = AccSpeciesName,
         new_try_name = scientificName) 

length(unique(try_wf_clean_second$scientificName)) #n = 753

# The species that do not have a direct match with the WFO backbone. n = 2323
to_be_matched <- not_included_in_wf %>%
  filter(AccSpeciesName %notin% c(second_check_matched$AccSpeciesName)) %>%
  pull(AccSpeciesName)


matched_with_wf <- WFO.match(spec.data = to_be_matched, WFO.data = classification)

save(try_filtered, try_clean2, removed_because_nonsensical, matched_with_wf,
 file = "data/processed/try_clean29_1.RData")

############ Maybe remove later
to_be_matched1 <- to_be_matched[1:500]
to_be_matched2 <- to_be_matched[501:1000]
to_be_matched3 <- to_be_matched[1001:1500]
to_be_matched4 <- to_be_matched[1501:2000]
to_be_matched5 <- to_be_matched[2001:2388]



#install.packages("fuzzyjoin")
#library(fuzzyjoin)

matched_dl_saved <- bind_rows(matched_with_wf_dl1, matched_with_wf_dl2, matched_with_wf_dl3,
                              matched_with_wf_dl4, matched_with_wf_dl5, matched_with_wf_dl6)

matched_with_wf_dl1 <- WFO.match.fuzzyjoin(spec.data = to_be_matched1, WFO.data = classification,
                                          stringdist.method = "dl")

matched_with_wf_dl2 <- WFO.match.fuzzyjoin(spec.data = to_be_matched2, WFO.data = classification,
                                          stringdist.method = "dl")

matched_with_wf_dl3 <- WFO.match.fuzzyjoin(spec.data = to_be_matched3, WFO.data = classification,
                                          stringdist.method = "dl")

matched_with_wf_dl4 <- WFO.match.fuzzyjoin(spec.data = to_be_matched4, WFO.data = classification,
                                          stringdist.method = "dl")

matched_with_wf_dl5 <- WFO.match.fuzzyjoin(spec.data = to_be_matched5, WFO.data = classification,
                                          stringdist.method = "dl")

matched_with_wf_dl_complete <- bind_rows(matched_with_wf_dl1, matched_with_wf_dl2,
                                         matched_with_wf_dl3, matched_with_wf_dl4,
                                         matched_with_wf_dl5) %>%
  relocate(scientificName, Old.status, Old.name, Fuzzy.dist) %>%
  filter(Fuzzy.dist <= 2) %>%
  group_by(spec.name.ORIG) %>%
  filter(n() > 1) %>%
  ungroup() %>%
  rename(scientificName_dl = scientificName,
         Old.name_dl = Old.name) %>%
  select(scientificName_dl, Old.name_dl, spec.name.ORIG)

matched_with_wf_save <- matched_with_wf %>%
  filter(Fuzzy.dist <= 2) %>%
  select(scientificName, spec.name.ORIG, Old.name, Fuzzy.dist)
  group_by(spec.name.ORIG) %>%
  filter(n() > 1) %>%
  ungroup() %>%
  rename(scientificName_l = scientificName,
         Old.name_l = Old.name) %>%
  select(scientificName_l, Old.name_l, spec.name.ORIG)

matched_compare_dist <- full_join(matched_with_wf_dl_complete, temp) %>%
  relocate(spec.name.ORIG)

difference <- matched_compare_dist %>%
  filter(spec.name.ORIG %in% c(to_be_matched)) %>%
  drop_na(scientificName_dl) %>%
  drop_na(scientificName_l) %>%
  filter(scientificName_l == scientificName_dl)
colSums(is.na(difference))
####### end of section maybe remove later


matched_with_wf_save <- matched_with_wf %>%
  filter(Fuzzy.dist <= 2) %>%
  select(scientificName, spec.name.ORIG, Old.name, Fuzzy.dist, taxonID) %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  group_by(spec.name.ORIG) %>%
  filter(n() == 1) %>% # Some names have multiple matches, they are removed
  ungroup()

length(unique(matched_with_wf_save$spec.name.ORIG)) #897
#write_xlsx(matched_with_wf_save, path = "matched_with_wf_save2.xlsx")

matched_with_wf_save_filt <- matched_with_wf_save %>%
  filter(spec.name.ORIG %notin% c("Brachyscome foliosa",
                                  "Buxus aemulans",
                                  "Capparis indica",
                                  "Ceropegia adenensis",
                                  "Ceropegia caudata",
                                  "Chenopodium wilsonii",
                                  "Clematis anethifolia",
                                  "Cleome maximiliani",
                                  "Clusia cretosa",
                                  "Cordia japonica",
                                  "Dacryodes canalensis",
                                  "Hibiscus americanus",
                                  "Lamium japonicum",
                                  "Limonium smithii",
                                  "Myriophyllum alternifolium",
                                  "Piper bellidi",
                                  "Pittosporum pullii",
                                  "Taxus grandis",
                                  "Urtica tenella",
                                  "Virola mollis",
                                  "Zygia media")) 

length(unique(matched_with_wf_save_filt$scientificName)) # 876 matches, 856 new wfo names

try_wf_clean_third <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(taxonID %in% c(matched_with_wf_save_filt$taxonID)) 
  
try_names_save_3 <- matched_with_wf_save_filt %>%
  select(spec.name.ORIG, scientificName) %>%
  rename(old_try_name = spec.name.ORIG,
         new_try_name = scientificName)


try_wf_clean_complete <- bind_rows(try_wf_clean_first, try_wf_clean_second,
                                   try_wf_clean_third) %>% distinct()

length(unique(try_wf_clean_complete$scientificName)) #65470

# Try_wf_clean2 contains the species in the TRY database that have a direct match
# within the WFO database. From that dataset, all names are removed that are 
# actually a synonym for another species. n = 
accepted_species_single_first <- try_wf_clean_complete %>% ## was eerst try_wf_clean2 
  filter(taxonomicStatus != "Synonym") %>%
  select(scientificName, taxonID, genus, family, taxonomicStatus) %>%
  rename(acceptedNameUsageID = taxonID, new_try_name = scientificName) %>%
  mutate(wfo_match = new_try_name) %>%
  group_by(wfo_match) %>% # Some species have multiple ID numbers, due
  # to a different associated author with the exact same name. First, the species 
  # are selected that that have no such duplicates
  filter(n() == 1) 

length(unique(accepted_species_single_first$wfo_match)) #58289

# These multiple accepted species names belonging to one TRY entry, poses a problem
# for the matching of species synonyms. As it is unclear of the synonym actually belongs to 
# the data in the TRY database. To check later in the script, if any of the matched
# accepted species names, could lead to misinterpretations, the AccSpeciesNames belonging
# to multiple WFO names, are saved. n = 1843. 
# There is a difference however, in the species that have multiple matches. Some of the
# matches are "Unchecked". And in some cases, if these are removed, only one "Accepted" match 
# remains. In that case, that match and the corresponding ID is selected and the species is
# not counted with the multiple matches list. 

accepted_species_single <- try_wf_clean_complete %>% #was eerst try_wf_clean2
  filter(taxonomicStatus != "Synonym") %>%
  select(scientificName, taxonID, genus, family, taxonomicStatus, nomenclaturalStatus) %>%
  rename(acceptedNameUsageID = taxonID, new_try_name = scientificName) %>%
  mutate(wfo_match = new_try_name) %>%
  group_by(wfo_match) %>%
  filter(n() > 1) %>% #2487
  ungroup() %>%
  filter(taxonomicStatus != "Unchecked")  %>%
  filter(taxonomicStatus != "Invalid")  %>% # no occurences
  distinct() %>%
  group_by(wfo_match) %>%
  filter(n()==1) %>% #1588 new matches
  bind_rows(accepted_species_single_first)

length(unique(accepted_species_single$wfo_match)) #59877

accepted_species_multiple <- try_wf_clean_complete %>% # was eerst try_wf_clean2
  filter(taxonomicStatus != "Synonym") %>%
  select(scientificName, taxonID, genus, family, 
         taxonomicStatus, nomenclaturalStatus) %>%
  rename(acceptedNameUsageID = taxonID, new_try_name = scientificName) %>%
  mutate(wfo_match = new_try_name) %>%
  filter(wfo_match %notin% c(accepted_species_single$wfo_match)) %>%
  arrange(acceptedNameUsageID) %>% # Some species have multiple ID numbers, due
  # to a different associated author with the exact same name. With the 
  # last lines of codes, the duplicates are removed. The 
  distinct(wfo_match, .keep_all = TRUE)

length(unique(accepted_species_multiple$wfo_match)) # n = 899
899+1588

accepted_species <- bind_rows(accepted_species_multiple, accepted_species_single) 

length(unique(accepted_species$new_try_name)) #60776


# The synonyms datafile contains all synonym names. Some of those names are the 
# same as an accepted species, indistinguishable if not for the author citation.
# These synonyms are removed as it is assumed that the name in the TRY database 
# corresponds with the accepted species name and not with the synonym. 

synonyms <- try_wf_clean_complete %>% # was eerst try_wf_clean2
  filter(taxonomicStatus == "Synonym") %>% # n = 8412
  filter(scientificName %notin% c(accepted_species$new_try_name)) %>%  # n = 3718
  # Filter out all the synonyms that have an actual species match. I am going to
  # assume that if an author writes down that name in the TRY database, 
  # it means the original, accepted species name and not a synonym of another species. 
  select(c(scientificName, acceptedNameUsageID, nomenclaturalStatus, 
           taxonomicStatus, genus, family)) %>%
  distinct()

length(unique(synonyms$scientificName)) # n = 4694


# First, the synonyms are extracted which have a unique, accepted name. To do so, first 
# the duplicate copies are removed (i.e. the speciesnames that occur more than once
# but still with the same acceptedNameUsageID. 
# In those cases, the author citation is different)

single_synonyms <- synonyms %>%
  select(c(scientificName, acceptedNameUsageID, genus, family)) %>%
  distinct() %>%
  group_by(scientificName, genus, family) %>%
  filter(n() == 1) %>%
  ungroup()

length(unique(single_synonyms$scientificName)) # n = 4495

#check <- multiple_synonyms %>%
 # rename(AccSpeciesName = scientificName) %>%
  #select(c(AccSpeciesName, acceptedNameUsageID)) %>%
  #left_join(codes) 

multiple_synonyms <- synonyms %>% 
  filter(scientificName %notin% c(single_synonyms$scientificName)) %>%
  #group_by(scientificName) %>%
  filter(nomenclaturalStatus %in% c(NA, "Valid")) # This also removes the synonym
 # Lupinus micranthus, as both and only matches are noted as "Illegitimate".


length(unique(multiple_synonyms$scientificName)) # n = 194

# These are the species names that are now resolved. n = 108
single_synonyms_2 <- multiple_synonyms %>%
  group_by(scientificName, genus, family ) %>%
  filter(n() == 1)

length(unique(single_synonyms_2$scientificName)) # n = 108

# These are the final synonym names that are resolved. n = 19
single_synonyms_3 <- multiple_synonyms %>%
  filter(scientificName %notin% c(single_synonyms_2$scientificName)) %>%
  group_by(scientificName, genus, family) %>%
  filter(any(!is.na(nomenclaturalStatus))) %>%
  drop_na() %>%
  filter(n() == 1)

length(unique(single_synonyms_3$scientificName)) # n = 19

# These are the synonyms that still have multiple, accepted species names
multiple_synonyms_final <- multiple_synonyms %>%
  filter(scientificName %notin% c(single_synonyms_2$scientificName)) %>%
  filter(scientificName %notin% c(single_synonyms_3$scientificName)) %>%
  distinct() %>%
  group_by(scientificName, genus, family) %>%
  
  filter(n() > 1) #67 synoniemen

length(unique(multiple_synonyms_final$scientificName)) # n = 67

check_names <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(taxonID %in% c(multiple_synonyms_final$acceptedNameUsageID)) %>%
  select(scientificName, taxonID) %>%
  distinct() %>%
  rename(acceptedNameUsageID = taxonID, new_name = scientificName) %>%
  full_join(multiple_synonyms_final)

# For some synonyms that still have multiple accepted species names, none of these
# accepted names are included in the TRY database. Note, these accepted names differ in 
# more than their author name. It is therefore save to keep those
# synonym names in the try database. 

multiple_synonyms_keep <- classification %>%
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  filter(taxonID %in% c(multiple_synonyms_final$acceptedNameUsageID)) %>%
  select(taxonID, scientificName, family) %>%
  rename(acceptedNameUsageID = taxonID, wfo_match = scientificName) %>%
  full_join(multiple_synonyms_final) %>%
  group_by(scientificName, family) %>%
  filter(!any(wfo_match %in% try_clean2$AccSpeciesName)) %>%
  ungroup() %>%
  select(scientificName, family) %>%
  rename(new_try_name = scientificName) %>%
  mutate(acceptedNameUsageID = "multiple",
         wfo_match = new_try_name) %>%
  separate(new_try_name, into = c("genus"), sep = " ", extra = "drop", remove = F) %>%
  distinct()

length(multiple_synonyms_keep$new_try_name)

all_synonyms <- bind_rows(single_synonyms, single_synonyms_2, single_synonyms_3) %>%
  mutate(include = "yes") %>%
  select(-c(genus, family))

how_many <- bind_rows(single_synonyms_2, single_synonyms_3)
length(unique(how_many$scientificName)) # n = 127

length(unique(all_synonyms$scientificName)) # n = 4622

# Now it is time to match the acceptedNameUsageID's with the actual species names
codes <- classification %>% 
  mutate(scientificName = str_replace_all(scientificName, "  ", " ")) %>%
  select(taxonID, scientificName, genus, family) %>%
  rename(acceptedNameUsageID = taxonID, wfo_match = scientificName) 


matched_synonyms <- full_join(all_synonyms, codes) %>%
  filter(include == "yes") %>%
  rename(new_try_name = scientificName) %>%
  filter(wfo_match %notin% c(accepted_species_multiple$wfo_match)) 
# n = 21 are removed

length(unique(matched_synonyms$new_try_name)) # 4601

# Some of the new names found by the synonyms, are also present in the
# accepted species list. This should not be a problem except for the fact that
# some species have the exact same genus and taxa name and are thus indistinguishable
# besides their author name which is not included in the TRY database. This leads to
# multiple species codes belonging to, for all for all intents and purposes, the same species.
# To make sure that I end up with a clean, clear dataset, where each new name belongs
# to only one species code, I check whether the matched synonym names also occur in the
# multiple accepted species list. If they do, they are removed. 


combined <- bind_rows(matched_synonyms, accepted_species) %>%
  select(-c(include, nomenclaturalStatus))

length(unique(combined$acceptedNameUsageID)) # HET KLOPT!! 64008 unique new species names and codes


# The combined dataset contains the accepted species name (the original try name),
# the new name according to the WFO, the corresponding WFO_id and taxonomic information
#try_direct_match2 <- try_clean2 %>%
 # filter(AccSpeciesName %in% c(try_wf_clean2$scientificName)) %>%
  #select(AccSpeciesName) %>%
  #distinct() %>%
  #full_join(combined) %>% #92 synonym species had multiple matches and where removed
  #drop_na(new_name) #

# This dataset contains new species names for the try names that did not have an
# initial match. The new names contain a mix of accepted and unchecked names. 
# No synonyms are included. If the new names (genus and epithet) are present more 
# than once in the classification dataset with the already matched try names,
# I'll have to remove them either way. Because then I have no way of knowing if the
# data belongs to the other data in the try database. 

#try_fuzzy_matched2 <- matched_with_wf %>%
 # relocate(Fuzzy.dist, spec.name.ORIG, scientificName, Subseq) %>%
#  filter(Matched == TRUE) %>%
 # filter(taxonRank %in% c("species", "subspecies", "variety")) %>%
  #filter(Fuzzy.dist <= 2) %>%
#  group_by(spec.name.ORIG) %>%
#  filter(n() == 1) %>%
#  ungroup() %>%
#  relocate(Old.name, Old.status) 
#  filter(scientificName %notin% c(accepted_species_multiple$AccSpeciesName)) %>%
#  select(-c(acceptedNameUsageID)) %>%
#  rename(AccSpeciesName = spec.name.ORIG, new_name = scientificName,
#         acceptedNameUsageID = taxonID) %>%
#  select(AccSpeciesName, new_name, genus, family, taxonomicStatus,
 #        acceptedNameUsageID) %>%
 # mutate(new_name = str_replace_all(new_name, "  ", " "))




#final <- bind_rows(try_direct_match2, try_fuzzy_matched)

final_new <- combined

combined2 <- final_new %>%
  select(-c(taxonomicStatus))


# 
final_combined <- combined2 %>%
  mutate(family =  case_when(genus == "Babbagia" ~  "Amaranthaceae",
                          genus ==  "Caelospermum" ~ "Rubiaceae",
                         genus == "Cimifuga" ~ "Ranunculaceae",
                         genus == "Daniella" ~ "Bignoniaceae",
                         genus == "Hablitzia" ~ "Amaranthaceae",
                         genus == "Leptogonum" ~ "Polygonaceae",
                         genus == "Namibithamnus" ~ "Asteraceae",
                         genus == "Nemum" ~ "Cyperaceae",
                         genus == "Petrophila" ~ "Proteaceae",
                         genus == "Platyspermation" ~ "Alseuosmiaceae",
                         genus == "Pseudopegolettia" ~ "Asteraceae",
                         genus == "Pterothamnus" ~ "Caprifoliaceae",
                         genus == "Szowitsia" ~ "Apiaceae",
                         genus == "Thylacium" ~ "Capparaceae",
                         genus == "Ardisia" ~ "Primulaceae", # for these next four, the genus was coupled to two families. 
                         genus == "Athyrium" ~ "Aspleniaceae",
                         genus == "Baeckea" ~ "Myrtaceae", 
                         genus == "Libertia" ~ "Iridaceae", # till here
                         genus == "Trigastrotheca" ~ "Molluginaceae",
                         genus == "Opoponax" ~ "Apiaceae",
                         TRUE ~ family)) %>%
  bind_rows(multiple_synonyms_keep)
  # genus == "Trigastrotheca" ~ "Molluginaceae"


try_tax <- vascular.families %>% # obtained from https://rdrr.io/cran/WorldFlora/man/vascular.families.html
  rename(family = Family, order = Order) %>%
  select(family, order, Group) %>%
  right_join(final_combined) %>%
  mutate(order = case_when(family == "Agdestidaceae" ~ "Caryophyllales",
                           family == "Aulacomniaceae" ~ "Bryales",
                           family == "Brachytheciaceae" ~ "Hypnales", 
                           family == "Bryaceae" ~ "Bryales", 
                           family == "Calymperaceae" ~ "Dicranales", 
                           family == "Daltoniaceae" ~ "Hookeriales", 
                           family == "Dicranaceae" ~ "Dicranales", 
                           family == "Didymochlaenaceae" ~ "Polypodiales", 
                           family == "Fissidentaceae" ~ "Dicranales", 
                           family == "Gymnomitriaceae" ~ "Jungermanniales", 
                           family == "Hemidictyaceae" ~ "Polypodiales", 
                           family == "Hookeriaceae" ~ "Hookeriales", 
                           family == "Hylocomiaceae" ~ "Hypnales", 
                           family == "Leucobryaceae" ~ "Dicranales", 
                           family == "Orthotrichaceae" ~ "Orthotrichales", 
                           family == "Pallaviciniaceae" ~ "Pallaviciniales", 
                           family == "Polytrichaceae" ~ "Polytrichales", 
                           family == "Pottiaceae" ~ "Pottiales", 
                           family == "Ptychomniaceae" ~ "Ptychomniales", 
                           family == "Pylaisiaceae" ~  "Hypnales", 
                           family == "Rhizogoniaceae" ~ "Rhizogoniales", 
                           family == "Ricciaceae" ~ "Marchantiales" , 
                           family == "Sematophyllaceae" ~ "Hypnales", 
                           family == "Sphagnaceae" ~ "Sphagnales", 
                           family == "Viburnaceae" ~ "Dipsacales",
                           family == "Cephalotaxaceae" ~ "Coniferales", 
                           family == "Amblystegiaceae" ~ "Hypnales",
                           family == "Calomniaceae"  ~ "Rhizogoniales",
                           family == "Corbichoniaceae" ~ "Caryophyllales", 
                           TRUE ~ order )) %>%
  filter(order %notin% c("Bryales", "Hookeriales", "Hypnales", "Dicranales", 
                         "Jungermanniales", "Marchantiales", 
                         "Orthotrichales", "Pallaviciniales", "Polytrichales",
                         "Pottiales", "Ptychomniales", "Rhizogoniales",
                         "Sphagnales")) %>%
  select(-c(Group))


try_names_save_complete <- bind_rows(try_names_save_1, try_names_save_2,
                                     try_names_save_3) 

try_tax2 <- try_tax %>% full_join(try_names_save_complete) %>% distinct() %>%
  rename(initial_wfo_match = new_try_name)

length(unique(try_tax2$old_try_name))

try_traits_tax_final <- try_clean2 %>%
  rename(old_try_name = AccSpeciesName) %>%
  right_join(try_tax2) %>%
  filter(LastName %notin% c("Frenette-Dussault")) %>%
  filter(Reference %notin% c("Jager et al (2015) Journal of Ecology 103(2):374-385. https://doi.org/10.1111/1365-2745.12366 ;  Simpson et al. (2016) Global Ecology and Biogeography 25:964-978 doi:10.1111/geb.12457", 
  "Raevel, V., Violle, C., & Munoz, F. (2012). Mechanisms of ecological succession: insights from plant functional strategies. Oikos, 121(11): 1761–1770. doi:10.1111/j.1600-0706.2012.20261.x Raevel, V., Munoz, F., Pons, V., Renaux, A., Martin, A., & Thompson, J.D. (2013). Changing assembly processes during a primary succession of plant communities on Mediterranean roadcuts. Journal of Plant Ecology, 6(1):19-28. https://doi.org/10.1093/jpe/rts011")) %>%
  drop_na()

save(try_traits_tax_final, file = "data.public/try_final291.RData")
length(unique(try_clean2$TraitName))


######

try_tax_hierarchy <- try_traits_tax_final %>%
  arrange(wfo_match) %>%
  select(wfo_match, genus, family, order) #%>%
  #distinct()

try_species_lvl <- try_traits_tax_final %>%
  select(wfo_match, StdValue, TraitName) %>%
 # group_by(wfo_match, TraitName) %>%
  summarize(across(where(is.numeric), ~mean(., na.rm = T))) %>%
  ungroup() %>%
  pivot_wider(names_from = TraitName, values_from = StdValue)

try_genus_lvl <- try_traits_tax_final %>%
  select(genus, StdValue, TraitName) %>%
  group_by(genus, TraitName) %>%
  summarize(across(where(is.numeric), ~mean(., na.rm = T))) %>%
  ungroup() %>%
  pivot_wider(names_from = TraitName, values_from = StdValue)

### Preparations for imputation
try_species_lvl_norm <- try_species_lvl %>%
  column_to_rownames(var = "wfo_match") %>%
  as.matrix()

#The data must be normalized and linearized. A log transformation is used. But it could be that that is not the appropriate transformation. 
back_trans_specieslvl <- list()
rm_col <- c()

for(i in 1:ncol(try_species_lvl_norm)){
x <- try_species_lvl_norm[,i] # goes through the columns
min_x <- min(x,na.rm = T) # takes the min of each column
if(min_x < 0.00000000001){
x <- x - min_x + 1 # make this optional if min x is neg
}
logx <- log10(x)
mlogx <- mean(logx, na.rm = T)
slogx <- sd(logx, na.rm = T)
x <- (logx - mlogx)/slogx # Z transformation
back_trans_specieslvl[[i]] <- list(min_x = min_x,
mlogx = mlogx,
slogx = slogx)
try_species_lvl_norm[,i] <- x
}


save(try_traits_tax_final, try_genus_lvl, try_species_lvl, try_tax_hierarchy, 
     try_species_lvl_norm, back_trans_specieslvl,
     file = "data/processed/try_final_clean161.RData")




#########@@@@@@@@@@@################## Einde poging 2 

```

```{r BHPMF}


trait_imp <- read.csv("data/processed/trait_imp.csv") %>%
  select(-c("X"))

## First, replace the permuted values with observed values where available, 
# then, scale variables to unit variance by dividing them with their sd

trait_perm_scaled <- trait_perm %>%
  rownames_to_column(var = "Name") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "value", names_to = "trait") %>%
  mutate(name_trait = paste(Name, trait)) %>%
  subset(name_trait %notin% trait_noNA$name_trait) %>%
  bind_rows(trait_noNA) %>%
  select(-c(name_trait)) %>%
  arrange(trait) %>%
  pivot_wider(names_from = "trait", values_from = "value") %>%
  mutate(across(where(is.numeric), ~ c(scale(., center = FALSE)))) %>%
  select(-c(BR_mean, BU_mean, LS_mean, RH_mean)) %>%
  arrange(Name)


## Check permuted values with observed values
### Do not run when generating model output

trait_perm <- trait_imp%>%
  select(-c(genus, family, order)) %>%
  relocate(wfo_match) %>%
  arrange(wfo_match) %>%
  rename("Leaf area (in case of compound leaves: leaflet, undefined if petiole is in- or excluded)" =
           "Leaf.area..in.case.of.compound.leaves..leaflet..undefined.if.petiole.is.in..or.excluded.",
         "Bark thickness" = "Bark.thickness",
         "Leaf carbon (C) content per leaf dry mass" = "Leaf.carbon..C..content.per.leaf.dry.mass",
         "Leaf carbon isotope signature (delta 13C)" = "Leaf.carbon.isotope.signature..delta.13C.",
         "Leaf chlorophyll content per leaf dry mass" = "Leaf.chlorophyll.content.per.leaf.dry.mass",
          "Leaf nitrogen (N) content per leaf dry mass" = "Leaf.nitrogen..N..content.per.leaf.dry.mass",
          "Leaf phosphorus (P) content per leaf dry mass" = "Leaf.phosphorus..P..content.per.leaf.dry.mass",
          "Leaf thickness" = "Leaf.thickness",
          "Plant height vegetative" = "Plant.height.vegetative",
          "Seed dry mass" = "Seed.dry.mass",
          "Plant lifespan (longevity)" = "Plant.lifespan..longevity.",
          "Leaf width" = "Leaf.width",
          "Leaf area per leaf dry mass (specific leaf area, SLA or 1/LMA): undefined if petiole is in- or excluded" =
           "Leaf.area.per.leaf.dry.mass..specific.leaf.area..SLA.or.1.LMA...undefined.if.petiole.is.in..or.excluded",
            "Leaf density (leaf tissue density, leaf dry mass per leaf volume)" = 
           "Leaf.density..leaf.tissue.density..leaf.dry.mass.per.leaf.volume.",
            "Leaf carbon/nitrogen (C/N) ratio" = "Leaf.carbon.nitrogen..C.N..ratio",
          "Leaf dry mass per leaf fresh mass (leaf dry matter content, LDMC)" = 
           "Leaf.dry.mass.per.leaf.fresh.mass..leaf.dry.matter.content..LDMC.",
          "Litter nitrogen (N) content per litter dry mass" = "Litter.nitrogen..N..content.per.litter.dry.mass",
          "Litter phosphorus (P) content per litter dry mass" = "Litter.phosphorus..P..content.per.litter.dry.mass",
          "Photosynthesis rate per leaf area" = "Photosynthesis.rate.per.leaf.area",
          "Seed length" = "Seed.length",
          "Stem specific density (SSD, stem dry mass per stem fresh volume) or wood density" = 
           "Stem.specific.density..SSD..stem.dry.mass.per.stem.fresh.volume..or.wood.density",
            "Leaf respiration rate in the dark per leaf area" = "Leaf.respiration.rate.in.the.dark.per.leaf.area",
          "Plant height" = "Plant.height",
          "Root diameter" = "Root.diameter",
          "Root porosity (fraction of root aerenchyma air space)" = "Root.porosity..fraction.of.root.aerenchyma.air.space.",
          "Root length per root dry mass (specific root length, SRL)" = "Root.length.per.root.dry.mass..specific.root.length..SRL.",
          "Root dry mass per root fresh mass (root dry matter content; RDMC)" = 
           "Root.dry.mass.per.root.fresh.mass..root.dry.matter.content..RDMC.",
      "Stem dry mass per stem fresh mass (stem dry matter content, StDMC)" = 
        "Stem.dry.mass.per.stem.fresh.mass..stem.dry.matter.content..StDMC.") %>%
  pivot_longer(cols = 2:ncol(.), names_to = "Trait",
               values_to = "Value_perm") 

trait_org <- try_species_lvl_norm %>%
  as.data.frame() %>%
  rownames_to_column(var = "wfo_match") %>%
  arrange(wfo_match) %>%
  pivot_longer(cols = 2:ncol(.), names_to = "Trait",
               values_to = "Value_org") %>%
  drop_na()

traits_std <- trait_org %>%
  group_by(Trait) %>%
  summarise(sd = sd(Value_org))

check_na <- try_traits_tax_final %>%
  select(c( ObservationID,  TraitName, StdValue)) %>%
  pivot_wider(values_from = StdValue, names_from = TraitName)

na_n <- (colSums(is.na(check_na))) %>%
  data.frame() %>%
  rename(na_n = ".") %>%
  mutate(perc_na = na_n/nrow(check_na)) %>%
  rownames_to_column(var = "Trait") %>%
  filter(Trait %notin% c("ObservationID"))

#write_xlsx(na_n, path = "traits_na_perc301.xlsx")

trait_both2 <- inner_join(trait_perm, trait_org) %>%
  mutate(SQ_ER = (Value_perm - Value_org)^2) %>%
  select(-c(Value_perm, Value_org)) %>%
  group_by(Trait) %>%
  summarise(MSE = mean(SQ_ER)) %>%
  ungroup() %>%
  mutate(RMSE = sqrt(MSE))


trait_both <- inner_join(trait_perm, trait_org) %>%
  mutate(SQ_ER = (abs(Value_perm - Value_org))) %>%
  select(-c(Value_perm, Value_org)) %>%
  group_by(Trait) %>%
  summarise(MSE = mean(SQ_ER))





```
