==> src/3. Complete_aiello_1992_V2.Rmd <==
---
title: "3. Complete analysis Eallonardo et al."
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Aiello-Lammens 
```{r clean data}





# The Aiello-lammens dataset contains 2 years. The first vegetation survey was done in 1992. The second one in 2011. In 2011, they did two plots per location to account for differences caused by slight differences in the location. I will analyse the data separately. 

aiello <- read_csv("data/raw/Aiello-Lammens et al. 2016/baviaanskloof_data_dryad.csv") %>%
  filter(Year == 1992) %>%
  rename(Name = Species, Plot_code = Plot,  cover = PercCover,
         Height = Height.Mean, LMA = LMA.Mean, Leaf_thickness = LeafThickness.Mean,
         Leaf_shape = LWR.Mean, FWC = FWC.Mean, FWC = FWC.Mean, twFWC = twFWC.Mean,
         Leaf_width = LeafWidthMax.Mean, Resprout = resprout_postfire) %>%
  mutate(Name = recode(Name, "Hypodiscus albo-aristatus" = "Hypodiscus alboaristatus",
                       "Pegolettia baccaridifolia" = "Pegolettia baccharidifolia",
                        "Diospyros austro-africana" = "Diospyros austroafricana",
                        "Rhodocoma capensis" = "Rhodocoma capense", 
                       "Erica passerinaee" = "Erica passerina" 
                       ))

aiello_veg <- aiello %>%
  select(Name, Plot_code, cover) %>%
  pivot_wider(names_from = Name, values_from = cover) %>%
  replace(is.na(.), 0) %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))

#aiello1992_species_list <- colnames(aiello_veg)

aiello_trait <- aiello %>%
  mutate(SLA = 1/LMA) %>%
  select(Name, Resprout, Height, SLA, Leaf_thickness, Leaf_shape, FWC,
         twFWC, Leaf_width) %>% distinct() %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) %>%
  drop_na()


aiello_trait_fd <- aiello_trait %>%
  column_to_rownames(var = "Name") 

species_list_fd <- aiello_trait %>% pull(Name)

aiello_veg_fd <- aiello_veg %>%
  select(all_of(species_list_fd))

aiello_trait <- aiello_trait %>%
  mutate(Resprout = recode(Resprout, `0` = 1,
                             `1`  =  2)) #%>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 
  
traitlist <- sort(c( "Resprout", "Height", "SLA", "Leaf_thickness", "Leaf_shape", 
                     "FWC", "twFWC", "Leaf_width"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

aiello1992_output <- lst(input_veg = aiello_veg, 
                         input_veg_fd = aiello_veg_fd,
                         input_trait = aiello_trait,
                         input_trait_fd = aiello_trait_fd,
                         cutoff_value = 0.8,
                         trait_detail = "Study",
                         trait_information = trait_information,
                         study_name = "Aiello1992")

save(aiello1992_output,
 file = "data/processed/aiello1992_prep.RData")



``` 


## 5. Perform calculations
```{r calculations }

aiello1992_processed_specieslvl <- cats_preparation(input_veg = aiello_veg, input_trait = aiello_trait, cutoff_value =  0.8, 
                                                 trait_information = trait_information, trait_detail = "Study")

mean_dist <- calculate_mean_dist(aiello_veg, aiello1992_processed_specieslvl)

FD <- dbFD(aiello_trait_fd, aiello_veg_fd, stand.x = T)

FD_df <- data.frame(FEve = FD$FEve,
        FDiv  = FD$FDiv,
        FRic = FD$FRic,
        RaoQ = FD$RaoQ,
        qual.FRic = FD$qual.FRic) %>%
  rownames_to_column(var = "Plot_code")
  

diversity_info <- aiello1992_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]] %>% 
  as.data.frame() %>% 
  mutate(shannon = diversity(.[,1:ncol(.)], index = "shannon", MARGIN = 1, base = exp(1)),
         simpson = diversity(.[,1:ncol(.)], index = "simpson", MARGIN = 1, base =exp(1)),
         species_richness = specnumber(.[,1:ncol(.)], MARGIN = 1),
         empty_states = ncol(aiello1992_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]]) - species_richness,
         meta_size = ncol(aiello1992_processed_specieslvl[["relative_abundances"]][["matrix4traits_prop"]])) %>%
  select(shannon, simpson, species_richness, empty_states, meta_size) %>% # base = exp(1) gives the shannon index a log e base
  rownames_to_column(var = "Plot_code") %>% 
  full_join(mean_dist) %>%
  full_join(FD_df)

```



## 6. Cats model
```{r cats calculations}

#Determines amount of cores, substract four to keep to pc free for other processes.
n_cores <- parallel::detectCores() - 5 

my_cluster <- makeSOCKcluster(n_cores)

registerDoSNOW(cl = my_cluster)

foreach::getDoParRegistered() # check if cluster is set up


aiello1992_cats <- cats_calculations(aiello1992_processed_specieslvl, trait_information, trait_detail = "Study") %>%
  mutate(study = "Aiello1992")

aiello1992_figure <- cats_visualisation(aiello1992_cats) %>%
  mutate(study = "Aiello1992")

save(aiello1992_processed_specieslvl, aiello1992_figure, aiello1992_species_list, aiello1992_cats,
 file = "data/processed/aiello1992_calculations.RData")

ggplot(aiello1992_figure, aes(x = species_richness, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between species richness and the CATS decomposition - Aiello-Lammers et al. 2016",
       x = "Species richness") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 

ggplot(aiello1992_figure, aes(x = mean_dist, y = KLR2)) +
  geom_point() +
  facet_wrap(~Deviance) + 
  theme_classic() +
  geom_smooth(method = "lm") + 
  labs(title = "The relationship between plot dissimilarity and the CATS decomposition - Aiello-Lammers et al. 2016",
       x = "Average Bray-Curtis index value") +
  geom_hline(yintercept = 0, linetype = 2) +
   coord_cartesian(ylim = c(-0.2, 1)) +
  scale_y_continuous(breaks = c(-0.2, 0,0.2, 0.4, 
    0.6,  0.8,  1
  )) 



load()


```
