==> src/3. Complete_asefa_V2.Rmd <==
---
title: "3. Complete analysis asefa et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 asefa et al. 2023
```{r clean data}

asefa_veg <- read_csv("data/raw/Asefa et al. 2023/species abundance.csv") %>%
  t() %>%
  as.data.frame() %>%
  rownames_to_column(var = "Name") %>%
  mutate(Name = gsub("_", " ", Name)) %>%
  mutate(Name = recode(Name, "Acacia abyssnica" = "Acacia abyssinica",
                       "Albizia gumifera" = "Albizia gummifera",
                       "Crateva adonsonia" = "Crateva adansonii",
                       "Olea europea subsp.cuspidata" = "Olea europaea subsp. cuspidata",
                       "Senna serguana" = "Senna singueana",
                       "Terminalia laxifolia" = "Terminalia laxiflora",
                       "Ziziphus spina" = "Ziziphus spina-christi")) %>%
  column_to_rownames(var = "Name") %>%
  t() %>%
  as.data.frame() %>%
  select(order(colnames(.)))

asefa_trait <- read_csv("data/raw/Asefa et al. 2023/trait data.csv") %>%
  rename(Name = species, Leaf_thickness = LT) %>%
  select(order(colnames(.))) %>%
  mutate(Name = gsub("_", " ", Name)) %>%
    mutate(Name = recode(Name, "Acacia abyssnica" = "Acacia abyssinica",
                       "Albizia gumifera" = "Albizia gummifera",
                       "Crateva adonsonia" = "Crateva adansonii",
                       "Olea europea subsp.cuspidata" = "Olea europaea subsp. cuspidata",
                       "Senna serguana" = "Senna singueana",
                       "Terminalia laxifolia" = "Terminalia laxiflora",
                       "Ziziphus spina" = "Ziziphus spina-christi")) %>%
  relocate(Name) %>%
  arrange(Name)


asefa_trait_fd <- asefa_trait %>%
  column_to_rownames(var = "Name") 

#asefa_trait <- asefa_trait %>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "LA", "LDMC", "Leaf_thickness", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)


asefa_output <- lst(input_veg = asefa_veg, 
                     input_veg_fd = asefa_veg,
                     input_trait = asefa_trait,
                     input_trait_fd = asefa_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Asefa")

save(asefa_output,
 file = "data/processed/asefa_prep.RData")

``` 


## 2. Perform calculations
```{r calculations }




########## later verwijderen
asefa_specieslvl <- complete_analysis(input_veg = asefa_veg, 
                            input_trait = asefa_trait,
                            input_trait_fd = asefa_trait_fd,
                            cutoff_value = 0.8,
                            trait_detail = "Study",
                            trait_information = trait_information,
                            study_name = "Asefa"
                            )


test <- complete_analysis(input_veg = asefa_output[["input_veg"]], 
                          input_veg_fd = asefa_output[["input_veg_fd"]],
                          input_trait = asefa_output[["input_trait"]],
                          input_trait_fd = asefa_output[["input_trait_fd"]],
                          cutoff_value = asefa_output[["cutoff_value"]],
                          trait_detail = asefa_output[["trait_detail"]],
                          trait_information = asefa_output[["trait_information"]],
                          study_name = asefa_output[["study_name"]])




```
