==> src/3. Complete_buzzard_V2.Rmd <==
---
title: "3. Complete analysis Buzzard et al. 2016"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Buzzard et al. 2016
```{r clean data}

buzzard <- read_csv("data/raw/Buzzard et al. 2015/FEBuzzardSpTraits.csv") %>%
  replace_na(list(species = "sp.")) %>%
  mutate(Name = paste0(genus, " ", species)) %>%
  rename(Plot_code = plot) %>%
  mutate(across(Plot_code, as.character)) %>%
  mutate(Name = recode(Name, "Allophyllusc occidentalisc" = "Allophylus occidentalis",
                       "Tabebuea ochracea" = "Tabebuia ochracea",
                       "Albizzia adinocephala" ="Albizia adinocephala",
                        "Karwinskia caldronii" = "Karwinskia calderonii",
                       "Dipterodendron costaricensis" = "Dipterodendron costaricense",
                       "Erythroxylon havanense" = "Erythroxylum havanense",
                       "Tabebuea rosea" = "Tabebuia rosea" 
                       ))

buzzard_veg <- buzzard %>%
  select(c(Plot_code, basal.area, Name)) %>%
  pivot_wider(values_from = basal.area, names_from = Name) %>%
  replace(is.na(.), 0) %>% 
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))

buzzard_veg_temp <- buzzard %>%
  select(c(Plot_code, basal.area, Name)) %>%
  pivot_wider(values_from = basal.area, names_from = Name) %>%
  replace(is.na(.), 0) %>%
  pivot_longer(cols = 2:ncol(.), values_to = "basal.area", names_to = "Name") 
  
  
buzzard_species_list <- colnames(buzzard_veg)
  
buzzard_trait <- buzzard %>%
  select(c(Plot_code, Name, SLA, LPC, CN, d13C)) %>%
  rename(Leaf_P = LPC, Leaf_CN = CN, D13C = d13C) %>%
  full_join(buzzard_veg_temp) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., mean(., na.rm = TRUE)))) %>%
  select(-c(basal.area, D13C)) %>%
  ungroup() %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  arrange(Name) %>%
  drop_na() #%>%
  #mutate(D13C = D13C+abs(min(D13C))) %>%
  #mutate(D13C = D13C+0.00000000001)


buzzard_trait_fd <- buzzard_trait

species_list_fd <- buzzard_trait %>% pull(Name)

buzzard_veg_fd <- buzzard_veg %>%
  select(all_of(species_list_fd))

#buzzard_trait <- buzzard_trait %>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 
 
traitlist <- sort(c("SLA", "Leaf_P", "Leaf_CN"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

buzzard_output <- lst(input_veg = buzzard_veg, 
                         input_veg_fd = buzzard_veg_fd,
                         input_trait = buzzard_trait,
                         input_trait_fd = buzzard_trait_fd,
                         cutoff_value = 0.8,
                         trait_detail = "Plot",
                         trait_information = trait_information,
                         study_name = "Buzzard")

save(buzzard_output,
 file = "data/processed/buzzard_prep.RData")


```
