==> src/3. Complete_castro_V2.Rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Castro et al. 2010
```{r clean data}

castro_species_list <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "splist") %>%
  mutate(Taxon = recode(Taxon, "Hypochoeris glabra" = "Hypochaeris glabra"))


plot_list <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "sitelist") %>%
    rename(Plot_code = Sites, 
           Sites = SiteID) %>%
    mutate(across(.cols = Sites, as.character))

castro_veg <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx") %>%
  drop_na() %>%
  column_to_rownames(var = "Sites")

colnames(castro_veg) <- castro_species_list$Taxon

castro_veg <- castro_veg %>%
  rownames_to_column(var = "Sites") %>%
  full_join(plot_list) %>%
  relocate(Plot_code) %>%
  separate(Plot_code, into = c("Treatment", "Replicate"), remove = F, sep = "(?<=\\D)(?=\\d)|(?<=\\d)(?=\\D)") 

castro_vegtemp <- castro_veg %>%
  select(-c(Sites, Replicate)) %>%
  pivot_longer(cols = 3:ncol(.), names_to = "Name", values_to = "cover")

castro_veg <- castro_veg %>%
  select(-c(Sites, Replicate, Treatment)) %>%
  column_to_rownames(var = "Plot_code") 


# They did not measure traits for each species in each species. Instead, they measured the traits only for the 80% most abundant ones. 

castro_trait <- read_excel("data/raw/CESTES/Plant_data/Castro2010_AJ.xlsx", 
    sheet = "traitsfull") %>%
  separate(Spraw, into = c("Name", "Treatment"), sep = "_") %>%
  rename(Height = VH, Leaf_N = LNC, Leaf_C = LCC) %>%
  select(-c(LF, M, Pollination, Dispersal, LPC, Spnew, Sp)) %>% # LPC removed, was not used in the original publication'
  mutate(Name = recode(Name, "Hypochoeris glabra" = "Hypochaeris glabra")) %>%
  full_join(castro_vegtemp) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., 
                        mean(., na.rm = TRUE)))) %>% # The NA's, caused by the absence of species in certain plots, are filled up by the species mean
  ungroup() %>%
  select(-c(cover, Treatment)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  arrange(Name) 

castro_trait_fd <- castro_trait 

# Standardizing is done after the fd dataframe is made, otherwise the values are standardized twice
#castro_trait <- castro_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


traitlist <- sort(c("Height", "Leaf_C", "LDMC", "Leaf_N", "SLA", "SM"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

castro_output <- lst(input_veg = castro_veg, 
                     input_veg_fd = castro_veg,
                     input_trait = castro_trait,
                     input_trait_fd = castro_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Treatment",
                     trait_information = trait_information,
                     study_name = "Castro")

save(castro_output,
 file = "data/processed/castro_prep.RData")


``` 


## 2. Cats model
```{r cats calculations}


castro_figure <- cats_visualisation(castro_cats) %>%
  separate(Plot_code, into = c("Treatment", "Replicate"), remove = F, sep = "(?<=\\D)(?=\\d)|(?<=\\d)(?=\\D)")
