==> src/3. Complete_chalmandrier_V2.rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")
```

## 1.Clean data

### 1.1 Chalmandrier et al. 2022
```{r clean data}

# The data of Chalmandrier has already been filtered in their own analysis. 
# I use their raw cover and trait data but I extract their selected plot_codes so I can filter the raw data

load("data/raw/Chalmandrier et al. 2022/data/data_ready.Rdata")

plot_code_vector <- env_p %>%
  pull(QuadID)

# Remove unnecessary objects from the global environment
rm(data)
rm(obs.comm)
rm(pdf_species)
rm(t.avg)
rm(env_p)

# Cleaning the vegetation data so it can be used to filter the trait data. 
# Some plots have traits but no cover data. The traits are used to impute missing values
# but the plots are afterwards removed. Some plots have cover but no traits, these traits
# are imputed by taking the species average across all plots. Some plots have cover and multiple
# trait values for some species. For those plots, the average species value is taken. In addition,
# trait data is created for the empty states,  ie. absent species in a certain plot. 

chalmandrier_species_list <- read_excel("data/raw/Chalmandrier et al. 2022/data/complete_species_names.xlsx")

chalmandrier_veg_cleantemp <- read_csv("data/raw/Chalmandrier et al. 2022/data/community.cover.23sp.csv") %>%
  filter(QuadID %in% plot_code_vector) %>%
  rename(Plot_code = QuadID) %>%
  pivot_longer(cols = 2:ncol(.), values_to = "cover", names_to = "Six-letter code") %>%
  left_join(chalmandrier_species_list) %>%
  select(-c(`Six-letter code`)) %>%
  rename(Name = Species)

chalmandrier_trait <- read_csv("data/raw/Chalmandrier et al. 2022/data/kettleholedata.23sp.csv") %>%
  filter(QuadID %in% plot_code_vector) %>%
  select(QuadID, poros, height, sla, species) %>%
  rename(Plot_code = QuadID, `Six-letter code` = species,
         Root_poros = poros, Height = height, SLA = sla) %>%
  left_join(chalmandrier_species_list) %>%
  rename(Name = Species) %>%
  select(-c(`Six-letter code`)) %>%
  group_by(Plot_code, Name) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>% #In some quadrants, some species have multiple recordings, the average is taken here. 
  ungroup() %>%
  full_join(chalmandrier_veg_cleantemp) %>%
  group_by(Name) %>%
  mutate(across(.cols = c(Root_poros, Height, SLA),  
            ~replace_na(., 
                        mean(., na.rm = TRUE)))) %>% # The NA's are filled up by the species mean
  ungroup() %>%
  arrange(Name) %>%
  drop_na(cover) %>%
  drop_na(Root_poros) %>% # Some cases have no abundance data and no trait data, by dropping NA's 
  # of one of the trait values, these cases (n = 134) are removed as well. 
  select(-c(cover)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name)

chalmandrier_trait_fd <- chalmandrier_trait 

# Standardizing is done after the fd dataframe is made, otherwise the values are standardized twice
#chalmandrier_trait <- chalmandrier_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


traitlist <- sort(c("Height", "Root_poros", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

# Cleaning the vegetation data so it can be used in the calculate_abundance function
chalmandrier_veg <- chalmandrier_veg_cleantemp %>%
  pivot_wider(names_from = "Name", values_from = "cover") %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))


chalmandrier_output <- lst(input_veg = chalmandrier_veg, 
                     input_veg_fd = chalmandrier_veg,
                     input_trait = chalmandrier_trait,
                     input_trait_fd = chalmandrier_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Plot",
                     trait_information = trait_information,
                     study_name = "Chalmandrier")

save(chalmandrier_output,
 file = "data/processed/chalmandrier_prep.RData")


```
