==> src/3. Complete_eallonardo_V2.Rmd <==
---
title: "3. Complete analysis Eallonardo et al."
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Eallonardo et al. 2012
```{r clean data}

eallonardo_veg_info <- read_excel("data/raw/CESTES/Plant_data/Eallonardo2013_AJ.xlsx", 
    sheet = "commfull") %>%
    rename(Plot_code = Sites_code) %>%
    select(c(Site, Plot_code, `T`))

eallonardo_species_list <- read_excel("data/raw/CESTES/Plant_data/Eallonardo2013_AJ.xlsx", 
  sheet = "splist") %>%
  rename(Sp = TaxCode, Name = FullName) %>%
  mutate(Name = recode(Name, "Asclepias syracia" = "Asclepias syriaca",
                       "Asparagus officicianalis" = "Asparagus officinalis",
                       "Aster laterflorus" = "Aster lateriflorus",
                       "Astriplex patula" = "Atriplex patula",
                       "Bohmeria cylindrica" = "Boehmeria cylindrica",
                       "Isis pseudacorus" = "Iris pseudacorus",
                       "Leptochlos fusca fascicularis" = "Leptochloa fusca subsp. fascicularis",
                       "Phragmites communis - native" = "Phragmites communis",
                       "Phragmites communis - non native" = "Phragmites communis",
                       "Physalis heterophyllum" = "Physalis heterophylla",
                       "Polygnum hydropiperoides" = "Polygonum hydropiperoides",
                       "Polygonum amphibinum" = "Polygonum amphibium",
                       "Sauruus cernuus" = "Saururus cernuus",
                       "Tecrium canadense" = "Teucrium canadense",
                       "Typha x glauca" = "Typha glauca"))


eallonardo_veg <- read_excel("data/raw/CESTES/Plant_data/Eallonardo2013_AJ.xlsx", 
    sheet = "commfull") %>%
    select(-c(`T`, P, NS, EW, Site, Sites_nr)) %>%
    pivot_longer(cols = 2:ncol(.), names_to = "Taxon", values_to = "cover") %>%
    left_join(eallonardo_species_list) %>%
    select(-c(Taxon, Sp)) %>%
    group_by(Name, Sites_code) %>%
    summarise(across(where(is.numeric), ~sum(.))) %>%
    ungroup() %>%
    pivot_wider(values_from = cover, names_from = Name) %>%
    column_to_rownames(var = "Sites_code" ) %>%
    select(order(colnames(.)))
    
relative_cover_check <- rowSums(eallonardo_veg) %>% as.data.frame()


eallonardo_trait <- read_excel("data/raw/CESTES/Plant_data/Eallonardo2013_AJ.xlsx", 
  sheet = "traitsfull")  %>%
  full_join(eallonardo_species_list) %>%
  rename(LA = Area_mm2, SLA = SLA_mm2_mg, Height = mature_ht_cm,
         Photo_path = C4, Leaf_CN = cn, Rhizome = Rhizomatous,
         Leaf_N = nmass, PLS = Perennial) %>%
  select(c(Name, LA, SLA, Height, Photo_path, PLS,
           Rhizome, Leaf_N, Leaf_CN)) %>%
  mutate(Photo_path = recode(Photo_path, `1` = 2,
                            `0`  =  1),
         PLS = recode(PLS, `1` = 2,
                             `0`  =  1),
         Rhizome = recode(Rhizome, `1` = 2,
                             `0`  =  1)) %>%
  group_by(Name) %>%
  summarise(across(where(is.numeric), ~mean(.))) %>%
  ungroup() %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) #%>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 


eallonardo_trait_fd <- read_excel("data/raw/CESTES/Plant_data/Eallonardo2013_AJ.xlsx", 
  sheet = "traitsfull")  %>%
  full_join(eallonardo_species_list) %>%
  rename(LA = Area_mm2, SLA = SLA_mm2_mg, Height = mature_ht_cm,
         Photo_path = C4, Leaf_CN = cn, Rhizome = Rhizomatous,
         Leaf_N = nmass, PLS = Perennial) %>%
  select(c(Name, LA, SLA, Height, Photo_path, PLS,
           Rhizome, Leaf_N, Leaf_CN)) %>%
  select(order(colnames(.))) %>%
  group_by(Name) %>%
  summarise(across(where(is.numeric), ~mean(.))) %>%
  ungroup() %>%
  relocate(Name) %>%
  arrange(Name) %>%
  column_to_rownames(var = "Name")

traitlist <- sort(c("Photo_path", "Leaf_CN", "Height", "LA", "Leaf_N",
                    "PLS", "Rhizome", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

eallonardo_output <- lst(input_veg = eallonardo_veg, 
                     input_veg_fd = eallonardo_veg,
                     input_trait = eallonardo_trait,
                     input_trait_fd = eallonardo_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Eallonardo")

save(eallonardo_output,
 file = "data/processed/eallonardo_prep.RData")


```
