==> src/3. Complete_gonzalez_V2_pa.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}

plot_meta <- read_csv("data/raw/González et al. 2021/bdj-09-e69560-s001.csv") %>%
  separate(locationID, into = c("site", "plot"), sep = 2, remove = F) %>%
  rename(Plot_code = locationID) %>%
  select(Plot_code, site)


gonzalez_species_list <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_scientific_name.csv")

removed_author <- gonzalez_species_list %>%
  separate(treeIdentification, into = c("genus", "sp"), sep = " ",
           extra = "merge", remove = F ) %>%
  filter(genus %notin% c("Unidentified")) %>%
  filter(sp %notin% c("sp. 1", "sp.", "sp. 2", "sp. 3",
                      "sp. 4", "sp. 6", "sp.2")) %>%
  separate(sp, into = c("sp2", "author"), extra = "merge", sep = " ",
           remove = F) %>%
  mutate(treeIdentification = paste0(genus, " ", sp2)) %>%
  select(treeIdentification, treeIdentificationID, family) %>%
  mutate(treeIdentification = recode(treeIdentification, "Diospyros digyna" = "Diospyros nigra",
                                     "Sapium nitidum" = "Sapium lateriflorum",
                                      "Trema micrantha" = "Trema micranthum"))
  # The synonym of Diospyros digyna Jacq. is Diospyros nigra 
 # (according to WFO: https://www.worldfloraonline.org/taxon/wfo-0000648775)

gonzalez_species_list <- gonzalez_species_list %>%
  filter(treeIdentificationID %notin% c(removed_author$treeIdentificationID)) %>%
  bind_rows(removed_author) %>%
  mutate(treeIdentificationID = recode(treeIdentificationID, "Drymis sp. 1" = "Drimys sp. 1",
                                       "Buddleia sp. 1" = "Buddleja sp. 1"))


gonzalez_veg <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_community_matrix.csv") %>%
  rename(Plot_code = locationID) %>%
  full_join(plot_meta) %>%
  filter(site == "Pa") %>%
  select(-site) %>%
  column_to_rownames(var = "Plot_code") %>%
  t() %>%
  as.data.frame() %>%
  rownames_to_column(var = "treeIdentificationID") %>%
  full_join(gonzalez_species_list) %>%
  column_to_rownames(var = "treeIdentification") %>%
  select(-c(treeIdentificationID, family)) %>%
  t() %>%
  as.data.frame() %>%
  select(order(colnames(.)))  

gonzalez_trait <- read_csv("data/raw/González et al. 2021/BIOVERA-tree_functional_traits.csv") %>%
  full_join(gonzalez_species_list) %>%
  select(c(treeIdentification, traitName, traitValue)) %>%
  rename(Name = treeIdentification) %>%
  group_by(Name, traitName) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  pivot_wider(values_from = traitValue, names_from = traitName) %>%
  drop_na() %>%
  rename(LA = leafArea, LDMC = leafDryMatterContent, Height = maximumHeight,
         SLA = specificLeafArea, WD = woodDensity, Leaf_thickness = leafThickness,
         Leaf_chlor = chlorophyllContent, Leaf_density = laminaDensity
         ) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name)

gonzalez_trait_fd <- gonzalez_trait %>%
  column_to_rownames(var = "Name") 

species_list_fd <- gonzalez_trait %>% pull(Name)

gonzalez_veg_fd <- gonzalez_veg %>%
  select(all_of(species_list_fd))

#gonzalez_trait <- gonzalez_trait %>%
#  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "Leaf_chlor", "Height", "LA", "Leaf_density", "LDMC",
                     "Leaf_thickness", "SLA", "WD"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

gonzalez_pa_output <- lst(input_veg = gonzalez_veg, 
                     input_veg_fd = gonzalez_veg_fd,
                     input_trait = gonzalez_trait,
                     input_trait_fd = gonzalez_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Gonzalez_pa")

save(gonzalez_pa_output,
 file = "data/processed/gonzalez_pa_prep.RData")

```
