==> src/3. Complete_jager_V2.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}

jager_trait <- read_csv("data/raw/Jager et al. 2014/speciestraits.csv")
coordinates <- read_csv("data/raw/Jager et al. 2014/plotenvironmentcoordinates.csv")

species_list <- read_excel("data/raw/Jager et al. 2014/speciescode_list.xlsx")


jager_veg <- read_csv("data/raw/Jager et al. 2014/speciesbasalarea.csv") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "BA", names_to = "species") %>%
  full_join(species_list) %>%
  select(-c(species)) %>%
  pivot_wider(values_from = BA, names_from = Name) %>%
  column_to_rownames(var = "plot") %>%
  select(order(colnames(.)))

jager_trait <- jager_trait %>%
  full_join(species_list) %>%
  rename(Leaf_thickness= leafthick, Leaf_density = leafden, SLA = sla, 
         LDMC = ldmc, WD = wood, Height = height, Bark_thickness = bark,
         SM = `seed mass`, Bark_thickness_rel = `relative bark thickness`,
         Leaf_N = leafn, Leaf_P = leafp, Litter_N = littern, 
         Litter_P = litterp, StDMC = tdmc) %>%
  select(-c(species)) %>%
  mutate(Bark_thickness = Bark_thickness + 0.00001,
         Bark_thickness_rel = Bark_thickness_rel + 0.0001) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name)

jager_trait_fd <- jager_trait %>%
  column_to_rownames(var = "Name") 

#jager_trait <- jager_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F))))  

jager_species_list <- colnames(jager_veg)

traitlist <- sort(c("Leaf_thickness", "Bark_thickness", "Bark_thickness_rel",
                    "SLA", "Leaf_density", "LDMC", "StDMC", "WD", 
                    "Leaf_N", "Leaf_P", "Litter_N", "Litter_P", "Height", "SM"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

jager_output <- lst(input_veg = jager_veg, 
                     input_veg_fd = jager_veg,
                     input_trait = jager_trait,
                     input_trait_fd = jager_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Jager")

save(jager_output,
 file = "data/processed/jager_prep.RData")

```
