==> src/3. Complete_jaschke_V2.Rmd <==
---
title: "3. Complete analysis Jäschke et al. 2019."
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Jäschke et al. 2019
```{r clean data}

jaschke <- read_excel("data/raw/Jäschke et al. 2019/PademosTibetTraitData-Dryad2019.10.25.xlsx", 
    skip = 1) %>%
  rename(Plot_code = Plot, cover = `Cover_%`, Name = Species,
         CN = `plant_C/N`, Height = CanoH_cm, SLA = `SLA_mm2/mg` ) %>%
  mutate(Name = recode(Name, "Elymus dura" = "Elymus durus",
                       "Ixeris chinensis ssp versicolor" = "Ixeris chinensis subsp. versicolor",
                       "Potentilla parvifollia" = "Potentilla parvifolia",
                       "Convolvulus amanii" = "Convolvulus ammannii",
                        "Caragana versicolor" = "Caragana versicolar"))

jaschke_veg <- jaschke %>%
  select(c(Plot_code, cover, Name)) %>%
  pivot_wider(values_from = cover, names_from = Name) %>%
  replace(is.na(.), 0) %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))


jaschke_veg_temp <- jaschke %>%
  select(c(Plot_code, cover, Name)) %>%
  pivot_wider(values_from = cover, names_from = Name) %>%
  replace(is.na(.), 0) %>%
  pivot_longer(cols = 2:ncol(.), values_to = "cover", names_to = "Name")
  
jaschke_species_list <- colnames(jaschke_veg)

jaschke_trait <- jaschke %>%
  select(c(Plot_code, Name, CN, Height, SLA)) %>%
  full_join(jaschke_veg_temp) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., mean(., na.rm = TRUE)))) %>%
  select(-c(cover)) %>%
  ungroup() %>%
  rename(Leaf_CN = CN) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  arrange(Name) 

jaschke_trait_fd <- jaschke_trait

#jaschke_trait <- jaschke_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 
  

traitlist <- sort(c("Leaf_CN", "Height", "SLA" ))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

jaschke_output <- lst(input_veg = jaschke_veg, 
                     input_veg_fd = jaschke_veg,
                     input_trait = jaschke_trait,
                     input_trait_fd = jaschke_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Plot",
                     trait_information = trait_information,
                     study_name = "Jäschke")

save(jaschke_output,
 file = "data/processed/jaschke_prep.RData")


```
