==> src/3. Complete_kober_V2.rmd <==
---
title: "0.Preprocessing"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Load data
```{r clean data}

kober_veg <- read_csv("data/raw/Kober et al. 2021/LocalAbundance.csv")

RegionalAbundance <- read_csv("data/raw/Kober et al. 2021/RegionalAbundance.csv")

traits <- read_csv("data/raw/Kober et al. 2021/traits.csv")

```

### 1.2 Preparation original data
```{r orginal data preparation}

# This is the original species list. Some species names are changed to make it comparable with
# the TRY database. 
kober_species_list <- RegionalAbundance %>%
  mutate(Species_long = recode(Species_long, `Achillea millefolium aggr.` = "Achillea millefolium",
                       `Medicago sativa aggr.` = "Medicago sativa",
                       `Arrhenaterum elatius` = "Arrhenatherum elatius",
                       "Senecio jacobea" = "Senecio jacobaea")) %>%
  select(-c(Abundance))

species_list <- kober_species_list %>% pull(Species_long)



# Some species have no presence in any of the plots. They are filtered out. 
# The original vegetation data contains species codes instead of actual species names
# Some original species name are not within the accepted species list but do have a match with the 
# SpeciesNames column 

kober_veg <- kober_veg %>% 
  select_if(~ !is.numeric(.) || sum(.) != 0) %>% # The code evaluates whether a column is numeric, if it is, the second statement is evaluated. 
  pivot_longer(cols = 2:ncol(.), names_to = "Species_short", values_to = "cover") %>%
  left_join(kober_species_list) %>%
  select(-c(Species_short)) %>%
  pivot_wider(names_from = Species_long, values_from = cover) %>%
  column_to_rownames(var = "plot") %>% 
  select(order(colnames(.))) 

# In the study
# The species Anthemis tinctoria has no trait data for the SLA. Instead, the family average is taken based on available data. 
traits_clean <- traits %>%
  select(c(site, spec_name, SLA, LDMC, Height)) %>%
  rename(Species_short = spec_name) %>%
  group_by(Species_short) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = T))) %>%
  ungroup() %>%
  mutate_all(~ifelse(is.nan(.), NA, .)) %>%
  left_join(kober_species_list) %>%
  rename(Name = Species_long) %>%
  select(-c(Species_short)) %>%
  arrange(Name) 

# To do so, the species list is extracted. This is used to pull up the hierarchical information using the lookup_table function.

species_list <- colnames(kober_veg)


hierarchy <- lookup_table(species_list, by_species=TRUE, 
                          family.tax = "plantlist", missing_action = "drop") %>%
  rownames_to_column(var = "Species_long") %>%
  arrange(Species_long) %>%
  rowid_to_column("plant_id") %>%
  rename(Name = Species_long) %>%
  full_join(traits_clean) %>%
  filter(Name %in% colnames(kober_veg))

family_mean <- hierarchy %>%
  select(c(Name, genus, family, SLA )) %>%
  group_by(family) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = T))) %>%
  filter(family == "Compositae") %>% # The familiy of the missing species
  pull(SLA) 

kober_trait <- hierarchy %>%
  select(Name, Height, LDMC, SLA) %>%
  replace_na(list(SLA = family_mean)) %>%
  arrange(Name) 

kober_trait_fd <- kober_trait %>%
  column_to_rownames(var = "Name") 

# Standardizing the traits is done after the trait_fd dataframe is created to avoid standardizing twice
#kober_trait <- kober_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- c("Height", "LDMC", "SLA")
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

kober_output <- lst(input_veg = kober_veg, 
                         input_veg_fd = kober_veg,
                         input_trait = kober_trait,
                         input_trait_fd = kober_trait_fd,
                         cutoff_value = 0.8,
                         trait_detail = "Study",
                         trait_information = trait_information,
                         study_name = "kober")

save(kober_output,
 file = "data/processed/kober_prep.RData")



```
