==> src/3. Complete_makelele_V2.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}

makelele <- read_csv("data/raw/Makelele et al. 2021/Yoko_chronosequence_transect.csv") %>%
  rename(Plot_code = plots, Name = species) %>%
  mutate(Name = recode(Name, "Quassia sylvestris" = "Quassia silvestris"))


makelele_trait <- makelele %>%
  select(Name, SLA, LDMC, C, N, P, WD) %>%
  rename(Leaf_C = C, Leaf_P = P, Leaf_N = N) %>%
  distinct() %>%
  group_by(Name) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  drop_na() %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name)

#makelele_veg <- makelele %>%
 # select(Plot_code, Name) %>%
  #group_by(Plot_code) %>%
  #count(Name) %>%
  #pivot_wider(names_from = Name, values_from = n) %>%
  #replace(is.na(.), 0) %>%
  #column_to_rownames(var = "Plot_code") %>%
  #select(order(colnames(.)))

makelele_veg <- makelele %>%
  select(Plot_code, Name, BA) %>%
  group_by(Plot_code, Name) %>%
  summarise(across(where(is.numeric), ~sum(., na.rm = T))) %>%
  pivot_wider(names_from = Name, values_from = BA) %>%
  replace(is.na(.), 0) %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))

makelele_trait_fd <- makelele_trait %>%
  column_to_rownames(var = "Name") 

species_list_fd <- makelele_trait %>% pull(Name)

makelele_veg_fd <- makelele_veg %>%
  select(all_of(species_list_fd))

makelele_species_list <- colnames(makelele_veg_fd)

#makelele_trait <- makelele_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "SLA", "LDMC", "Leaf_C", "Leaf_N", "Leaf_P", "WD"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

makelele_output <- lst(input_veg = makelele_veg, 
                     input_veg_fd = makelele_veg_fd,
                     input_trait = makelele_trait,
                     input_trait_fd = makelele_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Makelele")

save(makelele_output,
 file = "data/processed/makelele_prep.RData")

```
