==> src/3. Complete_osuri_V2_fragment.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}

osuri_veg <- read_csv("data/raw/Osuri et al. 2016/Osuri_Sanakran_2016_JAE_plot_data.csv") %>%
  filter(type == "FRAGMENT") %>%
  mutate(Plot_code = paste0(`plot.id`, `site.name`, type)) %>%
  select(Plot_code, species, `basal.area_sq.cm`) %>%
  group_by(Plot_code, species) %>%
  summarise(across(where(is.numeric), ~sum(., na.rm = T))) %>%
  ungroup() %>%
  mutate(species = recode(species, "Chrysophyllum roxburdhii" = "Chrysophyllum roxburghii",
                   "Goniothalamus cardiapetalus" = "Goniothalamus cardiopetalus",
                   "Otonephelum stipulaceum" = "Otonephelium stipulaceum",
                   "Reinwardtiodendron anaimalaiense" = "Reinwardtiodendron anamalaiense",
                   "Acronycchia pedunculata" = "Acronychia pedunculata")) %>%
  pivot_wider(values_from = `basal.area_sq.cm`, names_from = species) %>%
  replace(is.na(.), 0) %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))

osuri_trait <- read_excel("data/raw/Osuri et al. 2016/trait_data.xlsx") %>%
  rename(Name = Species, Seed_length = `Seed length (cm)`, SLA = `SLA (cm2.g-1)`,
         WD = `Wood density (g.cm-3)`, 
         Height = `Maximum attainable height (m)`) %>%
  select(-c(`No. of trees sampled`, Family, `Seed dispersal mode`)) %>%
  mutate(Seed_length = as.numeric(Seed_length),
         WD = as.numeric(WD)) %>% 
  #replace_na(list(WD = 1)) %>%
  drop_na() %>%
  select(order(colnames(.))) %>%
  mutate(Name = recode(Name, "Chrysophyllum roxburdhii" = "Chrysophyllum roxburghii",
                      "Goniothalamus cardiapetalus" = "Goniothalamus cardiopetalus",
                      "Otonephelum stipulaceum" = "Otonephelium stipulaceum",
                      "Reinwardtiodendron anaimalaiense" = "Reinwardtiodendron anamalaiense",
                      "Acronycchia pedunculata" = "Acronychia pedunculata")) %>%
  relocate(Name) %>%
  arrange(Name) %>%
  filter(Name %in% colnames(osuri_veg))
  

osuri_trait_fd <- osuri_trait %>%
  column_to_rownames(var = "Name") 

species_list_fd <- osuri_trait %>% pull(Name)

osuri_veg_fd <- osuri_veg %>%
  select(all_of(species_list_fd))

osuri_species_list <- colnames(osuri_veg_fd)

#osuri_trait <- osuri_trait %>%
#  mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c("Seed_length", "SLA", "WD", "Height"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

osuri_fragment_output <- lst(input_veg = osuri_veg, 
                     input_veg_fd = osuri_veg_fd,
                     input_trait = osuri_trait,
                     input_trait_fd = osuri_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Osuri_fragment")

save(osuri_fragment_output,
 file = "data/processed/osuri_fragment_prep.RData")

```
