==> src/3. Complete_raevel_V2.Rmd <==
---
title: "3. Complete analysis Raevel et al. 2012"
author: "Jelyn Gerkema"
date: "2023-06-20"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Raevel et al. 2012
```{r clean data}

raevel_veg <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
    sheet = "comm") %>%
    column_to_rownames(var = "Sites") 

raevel_species_list <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
    sheet = "splist") %>%
  rename(Sp = TaxCode, Name = TaxName) %>%
   mutate(Name = paste0(tools::toTitleCase(word(Name, 1)), " ", word(Name, 2, -1))) %>%
  mutate(Name = recode(Name, "Dianthus longicaulis" = "Dianthus sylvestris subsp. longicaulis",
                       "Scabiosa atropurpurea var. maritima" = "Scabiosa atropurpurea" ))

# Based on their author name. 

colnames(raevel_veg) <- raevel_species_list$Name

raevel_veg <- raevel_veg %>%
  select(order(colnames(.)))

    colnames(raevel_veg)
raevel_trait <- read_excel("data/raw/CESTES/Plant_data/Raevel2012_AJ.xlsx", 
  sheet = "traits") %>%
  full_join(raevel_species_list) %>% 
  rename(Height = H ) %>%
  select(-c(Sp, DM, LS, LF, FS,  CodeOri, Authorities, ScientificName, Family, Rank)) %>%
  select(order(colnames(.))) %>%
  relocate(Name) %>%
  arrange(Name) 

raevel_trait_fd <- raevel_trait %>%
  column_to_rownames(var = "Name")

#raevel_trait <- raevel_trait %>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = FALSE)))) 



traitlist <- sort(c( "Height", "SLA", "SM"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

raevel_output <- lst(input_veg = raevel_veg, 
                     input_veg_fd = raevel_veg,
                     input_trait = raevel_trait,
                     input_trait_fd = raevel_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Raevel")

save(raevel_output,
 file = "data/processed/raevel_prep.RData")


```
