==> src/3. Complete_rolhauser_V2.Rmd <==
---
title: "3. Complete analysis Rolhauser et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 Rolhauser et al. 2021
```{r clean data}

species_codes <- read_excel("data/raw/Rolhauser et al. 2021/species_codes_rolhauser.xlsx", 
    sheet = "codes_fullnames") %>%
  separate(codes, into = c("sp", "Name"), sep = " - ") %>%
  mutate(Name = recode(Name, "Polygonum biflorum" = "Polygonatum biflorum"))

# Rolhauser collected frequency data. They had 189 sites, and for each site, they recorded species presence in multiple subplots. The amount of subplots howerever, differs per site. To make the plots comparable with each other, the frequency is divided by the total number of subplots. 
rolhauser_veg <- read_excel("data/raw/Rolhauser et al. 2021/Copy of Rolhauser__Waller___Tucker_J_Ecol_-_data.xlsx") %>%
  mutate(freq_new = freq/quads) %>%
  select(c(site, sp, freq_new)) %>%
  full_join(species_codes) %>%
  select(-c(sp)) %>%
  pivot_wider(names_from = Name, values_from = freq_new) %>%
  rename(Plot_code = site) %>%
  column_to_rownames(var = "Plot_code") %>%
  select(order(colnames(.)))

rolhauser_trait <- read_excel("data/raw/Rolhauser et al. 2021/Copy of Rolhauser__Waller___Tucker_J_Ecol_-_data.xlsx") %>%
  select(c(sp, SLA, VH, LCC, LS, LT, LNC, LDMC)) %>%
  distinct() %>%
  full_join(species_codes) %>%
  select(-c(sp)) %>%
  rename(Height = VH, Leaf_C = LCC, Leaf_N = LNC,
         LA = LS, Leaf_thickness = LT) %>%
  select(-Leaf_N, LA, Leaf_thickness, LDMC) %>%
  arrange(Name) %>%
  select(order(colnames(.))) %>%
  relocate(Name) #%>%
  #mutate(across(where(is.numeric), ~ c(scale(., center = F))))

rolhauser_trait_fd <- rolhauser_trait %>%
  column_to_rownames(var = "Name")  # Center is true, this results in a mean of 0. This way of standardizing would not work for CATS, since all values need to be higher than 0, but for FD, this is preferred. 
  
# "LDMC", "Leaf_N", "LA", "Leaf_thickness",
traitlist <- sort(c( "Height", "Leaf_C",   "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

rolhauser_output <- lst(input_veg = rolhauser_veg, 
                     input_veg_fd = rolhauser_veg,
                     input_trait = rolhauser_trait,
                     input_trait_fd = rolhauser_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Study",
                     trait_information = trait_information,
                     study_name = "Rolhauser")

save(rolhauser_output,
 file = "data/processed/rolhauser_prep.RData")

```
