==> src/3. Complete_standen_V2.Rmd <==
---
title: "3. Complete analysis Kassaw et al."
author: "Jelyn Gerkema"
date: "2023-10-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## 0. Set up
```{r set up}
source("src/settings2.R")
source("src/functions.R")

```

## 1.Clean data


### 1.1 
```{r clean data}


# dataset consists of 40 plots with two subplot each. In the original article, the community is averages for the two subplots.
# I have done the same here
standen_veg_species_list <- read_excel("data/raw/Standen et al. 2021/Community_Composition_Scotty Creek_Standen 2021_RAW.xlsx", 
    sheet = "Meta Data") %>% 
  select(c(`Species Abbreviations`, `Species List (Genus species)`))

standen_veg <- read_excel("data/raw/Standen et al. 2021/Community_Composition_Scotty Creek_Standen 2021_RAW.xlsx") %>%
  pivot_longer(cols = 5:ncol(.), names_to = "Species Abbreviations", values_to = "abundance") %>%
  group_by(Plot, `Species Abbreviations`) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  full_join(standen_veg_species_list) %>%
  select(-c(`Species Abbreviations`, Quadrat)) %>%
  pivot_wider(names_from = `Species List (Genus species)`, values_from = abundance) %>%
  column_to_rownames(var = "Plot") %>%
  select(order(colnames(.)))  

standen_veg_temp <- standen_veg %>%
  rownames_to_column(var = "Plot_code") %>%
  pivot_longer(cols = 2:ncol(.), values_to = "cover", names_to = "Name")

standen_trait_species_list <- read_excel("data/raw/Standen et al. 2021/Trait Data_Scotty Creek_Standen2021.xlsx", 
    sheet = "Meta Data") %>%
    filter(Variable == "Species") %>%
    separate(Notes, into = c(as.character(1:18)), sep = ";") %>%
    select(-c(Variable, Definition, Units)) %>%
    t() %>%
    as.data.frame() %>%
    separate(V1, into = c("Species", "Name"), sep = "=") %>%
    mutate(across(where(is.character), str_trim))
    

standen_trait <- read_excel("data/raw/Standen et al. 2021/Trait Data_Scotty Creek_Standen2021.xlsx") %>%
  full_join(standen_trait_species_list) %>%
  rename(Plot_code = Plot, SLA = SLA_mm2_mg, Leaf_N = N_mass_mg_g, 
         Photo_rate = A_mass_umol_g_s, Leaf_resp = R_mass_umol_g_s) %>%
  group_by(Plot_code, Name) %>%
  summarise(across(where(is.numeric), ~mean(., na.rm = TRUE))) %>%
  ungroup() %>%
  full_join(standen_veg_temp) %>%
  drop_na(cover) %>%
  group_by(Name) %>%
  mutate(across(where(is.numeric), ~replace_na(., mean(., na.rm = TRUE)))) %>%
  ungroup() %>%
  drop_na() %>%
  select(-c(cover)) %>%
  select(order(colnames(.))) %>%
  relocate(Plot_code, Name) %>%
  arrange(Name)

standen_trait_fd <- standen_trait

species_list_fd <- standen_trait %>% pull(Name)

standen_veg_fd <- standen_veg %>%
  select(all_of(species_list_fd))

standen_species_list <- colnames(standen_veg_fd)

#standen_trait <- standen_trait %>%
 # mutate(across(where(is.numeric), ~ c(scale(., center = F)))) 

traitlist <- sort(c( "Leaf_N", "Photo_rate", "Leaf_resp", "SLA"))
cwm_names <- paste0(traitlist, "_cwm")
cwv_names <- paste0(traitlist, "_cwv")

trait_information <- lst(traitlist, cwm_names, cwv_names)

standen_output <- lst(input_veg = standen_veg, 
                     input_veg_fd = standen_veg_fd,
                     input_trait = standen_trait,
                     input_trait_fd = standen_trait_fd,
                     cutoff_value = 0.8,
                     trait_detail = "Plot",
                     trait_information = trait_information,
                     study_name = "Standen")

save(standen_output,
 file = "data/processed/standen_prep.RData")

```
