==> src/5. Simulation.Rmd <==
---
title: "Simulations"
author: "Jelyn Gerkema"
date: "2024-03-13"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r simulations}


test <- cats_sim(lambda=0.3,meta.community.u=0.3,random.weight=0.2,Nperm=500,Nsims=1,
                     association="yes",sd.metacommunity.x=1)

lambda=0.3
meta.community.u=0.3
random.weight=0.2
                   Nperm=500
Nsims=10
                     association="yes"
sd.metacommunity.x=1


o <- test$observed.ra

m <- test$observed.metacommunity.ra


cor(o[1,1:10], m[1,1:10])

test[1:4]

com <- data.frame(o = o[1,1:10], m = m[1, 1:10], trait = t(test$trait)) %>%
  rownames_to_column(var = "species")

ggplot(com, aes(x = o, y = m, colour = species)) +
  geom_point()

CWM<-ra%*%t(trait)

test[15]

```
